Updated to use refactored UDT photo uploading bottom sheet class
- PhotoUploadBottomSheetDialog has been migrated to become the parent class in which a child class should be defined to use it
